review: blackbox follow-ups (flags comment, PsramSink non-copyable, doc gate, test coverage)
- controller.cpp: 注释说明在控路径 link/imu 标志恒置位
- board_config.h: 记录 WP_HAS_BLACKBOX 启用路径
- PsramSink: 删除拷贝/移动构造，防双重释放
- frame test: 补齐 gyro_y/accel_x/accel_y/pitch_deg round-trip 断言
- sink test: 新增环回字节摆放验证
- getU16: 显式 cast 与 getU32 一致

+++ core/blackbox/blackbox_frame.cpp
@@ -14,7 +14,9 @@ uint32_t getU32(const uint8_t*& p) {
     uint32_t v = uint32_t(p[0]) | (uint32_t(p[1]) << 8) |
                  (uint32_t(p[2]) << 16) | (uint32_t(p[3]) << 24); p += 4; return v;
 }
-uint16_t getU16(const uint8_t*& p) { uint16_t v = uint16_t(p[0] | (p[1] << 8)); p += 2; return v; }
+uint16_t getU16(const uint8_t*& p) {
+    uint16_t v = uint16_t(uint16_t(p[0]) | (uint16_t(p[1]) << 8)); p += 2; return v;
+}
 float getF(const uint8_t*& p) { uint32_t u = getU32(p); float v; std::memcpy(&v, &u, 4); return v; }
 }  // namespace
